refactor: split crypto functions out from util/util to util/crypto.
* update references, move tests.
