+++ lib/model/instance/session.js
@@ -1,7 +1,7 @@
 const { withCreateTime } = require('../../util/instance');
 const Instance = require('./instance');
 const Option = require('../../reused/option');
-const { generateToken } = require('../../util/util');
+const { generateToken } = require('../../util/crypto');
 
 
 const sessionFields = [ 'actorId', 'token', 'expiresAt', 'createdAt' ];
