+++ lib/model/instance/user.js
@@ -1,6 +1,7 @@
 const Instance = require('./instance');
 const { merge, mergeAll, pick } = require('ramda');
-const { hashPassword, isBlank } = require('../../util/util');
+const { isBlank } = require('../../util/util');
+const { hashPassword } = require('../../util/crypto');
 
 const userFields = [ 'actorId', 'password', 'mfaSecret', 'email' ];
 Object.freeze(userFields);
