+++ lib/resources/sessions.js
@@ -1,5 +1,6 @@
 const Problem = require('../problem');
-const { isBlank, verifyPassword } = require('../util/util');
+const { isBlank } = require('../util/util');
+const { verifyPassword } = require('../util/crypto');
 const { endpoint, getOrReject, success } = require('../util/http');
 
 module.exports = (service, { User, Session }) => {
