+++ lib/resources/users.js
@@ -1,5 +1,5 @@
 const { endpoint, getOrNotFound, success, isTrue } = require('../util/http');
-// const { verifyPassword } = require('../util/util');
+// const { verifyPassword } = require('../util/crypto');
 const Problem = require('../problem');
 const { resolve, reject } = require('../reused/promise');
