+++ lib/util/util.js
@@ -1,7 +1,4 @@
 const { inspect } = require('util');
-const { ExplicitPromise, resolve } = require('../reused/promise');
-const bcrypt = require('bcrypt');
-const { randomBytes } = require('crypto');
 
 const isBlank = (x) => (x === null) || (x === undefined) || (x === '');
 
@@ -39,15 +36,5 @@ const incr = () => {
   };
 };
 
-const hashPassword = (plain) =>
-  (isBlank(plain) ? resolve(null) : ExplicitPromise.of(bcrypt.hash(plain, 12)));
-const verifyPassword = (plain, hash) => ((isBlank(plain) || (isBlank(hash)))
-  ? resolve(false)
-  : ExplicitPromise.of(bcrypt.compare(plain, hash)));
-
-const generateToken = () => randomBytes(48).toString('base64')
-  .replace(/\//g, '!')
-  .replace(/\+/g, '$');
-
-module.exports = { isBlank, printPairs, without, get, ensureArray, incr, hashPassword, verifyPassword, generateToken };
+module.exports = { isBlank, printPairs, without, get, ensureArray, incr };
